+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+- Support `Serialized Type` to Script txn.  Now can use vector<String> for example.
 - Add optional address parameter to MultiKeyAccount constructor
 
 # 1.27.1 (2024-08-23)